+++ CHANGELOG.md
@@ -15,6 +15,7 @@
  *   Added intructions to get date and mouse/touch/keyboard events
  *   Allow gridStep to go down to 1
  *   More web app devices compatibility
+ *   Fix: Copy&paste in Windows now keeps newlines (thank you Jordi Losantos)
  *   Fix: Save code now works with all browsers (except Safari, where it will pop a window)
  *   Fix: jsColor iputs missed the '#' at the beginning
  *   Fix: Added jscolor support for windowInputEdit/Create which broke in IE